+++ CHANGELOG.md
@@ -5,6 +5,8 @@
   ## **WORK IN PROGRESS**
   * (theimo1221) Update packages
 -->
+## **WORK IN PROGRESS**
+* (theimo1221) Make camera video optional
 ## 0.5.3 (2024-12-31)
 * (theimo1221) Some changes to audio and experimental HKSV support
 * (theimo1221) Allow option to skip certain devices for homebridge